backend/requirements.txt
Django>=5.2
djangorestframework>=3.15
django-cors-headers>=4.3

psycopg2-binary>=2.9
